Fix more linting and refactor
